requirements.txt
requests
pdfplumber
python-dateutil
python-dotenv
